Updated git submodules
Project: why3 master e08df79fbe29779678436e82051114083370f81b

S311-005 Filter model for SPARK counterexamples

This removes any counterexamples names that do not begins with a number or
a period (to reduce size of the output). Those are not parsed by gnat2why
anyway.

* src/core/model_parser.ml*
(spark_filter_model): Add a specific filtering of model for SPARK (not to
be merged in Why3).

* src/gnat/gnat_report.ml
(print_cntexmp_model): First filter model before printing it

Change-Id: I52c3e698f0f1df34b214480bda742bafe2dc1550